Base.convert

Base.convert

ParSitter.build_tq_tree

ParSitter.build_xml_tree

ParSitter.check_tq_tree

ParSitter.get_capture

ParSitter.is_capture_node

ParSitter.match_tree

ParSitter.query

Base.convert — Method

Convert a tree t to a TreeQueryExpr. The node value is returned by nodevalue and children of the node returned by children.

Base.convert — Method

Convert from a tree-like object to a Tuple.

using ParSitter
tt=(1,2,(3,(4,5,(6,),7,5)));
tq = ParSitter.build_tq_tree(tt)
tt2 = convert(Tuple, tq)

ParSitter.build_tq_tree — Method

build_tq_tree(t::Tuple)

Build a tree query expression tree out of a nested tuple: the assumption is that the first element of each tuple is the head of the expression, the rest are children.

ParSitter.build_xml_tree — Method

build_xml_tree(tree_sitter_xml_ast::String)

Postprocesses and parse the tree-sitter XML output to something that can be traversed to match a given query.

ParSitter.check_tq_tree — Method

Checks that a query tree does not contain duplicate capture keys.

ParSitter.get_capture — Method

Function that returns captured values from a query. If the key does not
exist `nothing` is returned.

ParSitter.is_capture_node — Method

is_capture_node(n; capture_sym=DEFAULT_CAPTURE_SYM)

Function that checks whether a node is a 'capture node' i.e. value of the form "match@capture_key" and returns a NamedTuple with the result and the capture key string

julia> ParSitter.is_capture_node("value@capture_key")
(is_match = true, capture_key = "capture_key")

julia> ParSitter.is_capture_node("value@capture_key", capture_sym="@@")
(is_match = false, capture_key = nothing)

ParSitter.match_tree — Method

function match_tree(target_tree,
                    query_tree;
                    match_cache=Dict(),
                    captured_symbols=MultiDict(),
                    match_type=:strict,
                    is_capture_node=is_capture_node,
                    target_tree_nodevalue=AbstractTrees.nodevalue,
                    query_tree_nodevalue=AbstractTrees.nodevalue,
                    capture_function=AbstractTrees.nodevalue,
                    node_comparison_yields_true=(args...)->false)

Function that searches a query_tree into a target_tree. It returns a vector of subtree matches, where each element is a Tuple that contains the result of the match, any captured values and the trees that were compared. To capture a value, the function is_capture_node must return true for a given query node. One example is using query nodes of the form "nodevalue@capture_variable". In the matching process, the query and target node values are extracted using query_tree_nodevalue and target_tree_nodevalue respectively and compared. If they match, the target_tree node value is captured by applying capture_function to the node and a MultiDict("capture_variable"=>captured_target_node_value)). Tree comparisons are also hashed and the result as well as captured symbols are stored in match_cache for quick retrieval.

Example

julia> using ParSitter
       using AbstractTrees

       _query_tree_nodevalue(n) = ParSitter.is_capture_node(n).is_match ? split(n.head, "@")[1] : n.head
       _target_tree_nodevalue(n)=string(n.head)
       _capture_on_empty_query_value(t1,t2) = ParSitter.is_capture_node(t2).is_match && isempty(_query_tree_nodevalue(t2))

       my_matcher(t,q) = ParSitter.match_tree(
                              ParSitter.build_tq_tree(t),
                              ParSitter.build_tq_tree(q);
                              target_tree_nodevalue=_target_tree_nodevalue,
                              query_tree_nodevalue=_query_tree_nodevalue,
                              capture_function=n->n.head,
                              node_comparison_yields_true=_capture_on_empty_query_value)

       query = ("1@v0", "2", "@v2")   # - query means: capture in "v0" if target value is 1, match on 2, capture any symbol in "v2"

       t=(1,2,10); my_matcher( t, query)[1:2] |> println
       t=(10,2,11); my_matcher( t, query)[1:2] |> println
       t=(1,2,3,4,5); my_matcher( t, query)[1:2] |> println
(true, MultiDict{Any, Any}(Dict{Any, Vector{Any}}("v2" => [10], "v0" => [1])))
(false, MultiDict{Any, Any}(Dict{Any, Vector{Any}}("v2" => [11])))
(true, MultiDict{Any, Any}(Dict{Any, Vector{Any}}("v2" => [3], "v0" => [1])))

ParSitter.query — Method

Query a tree with another tree. This will match the query_tree with all substrees of target_tree. Both trees should support the AbstractTrees interface.

Example

julia> using ParSitter
       using AbstractTrees

       query = ("1@v0", "2", "@v2")   # - query means: capture in "v0" if target value is 1, match on 2, capture any symbol in "v2"
       target = (1, 2, 3, (10, 2, 3)) # - only the (1,2,3) subtree will match, the second will not bevause of the 10;
                                      # - @v2 will always capture values (due to `_capture_on_empty_query_value`)
       query_tq = ParSitter.build_tq_tree(query)
       target_tq = ParSitter.build_tq_tree(target)

       _query_tree_nodevalue(n) = ParSitter.is_capture_node(n).is_match ? split(n.head, "@")[1] : n.head
       _target_tree_nodevalue(n) = string(n.head)
       _capture_on_empty_query_value(t1,t2) = ParSitter.is_capture_node(t2).is_match && isempty(_query_tree_nodevalue(t2))
       print_tree(target_tq); println("---")
       print_tree(query_tq); println("---")
       r=ParSitter.query(target_tq,
                         query_tq;
                         match_type=:strict,
                         target_tree_nodevalue=_target_tree_nodevalue,
                         query_tree_nodevalue=_query_tree_nodevalue,
                         capture_function=n->n.head,
                         node_comparison_yields_true=_capture_on_empty_query_value)
       map(t->t[1:2], r)
1
├─ 2
├─ 3
└─ 10
   ├─ 2
   └─ 3
---
"1@v0"
├─ "2"
└─ "@v2"
---
6-element Vector{Tuple{Bool, MultiDict{Any, Any}}}:
 (1, MultiDict{Any, Any}(Dict{Any, Vector{Any}}("v2" => [3], "v0" => [1])))
 (0, MultiDict{Any, Any}(Dict{Any, Vector{Any}}()))
 (0, MultiDict{Any, Any}(Dict{Any, Vector{Any}}()))
 (0, MultiDict{Any, Any}(Dict{Any, Vector{Any}}("v2" => [3])))
 (0, MultiDict{Any, Any}(Dict{Any, Vector{Any}}()))
 (0, MultiDict{Any, Any}(Dict{Any, Vector{Any}}()))

julia> r=ParSitter.query(target_tq,
                         query_tq;
                         match_type=:nonstrict,
                         target_tree_nodevalue=_target_tree_nodevalue,
                         query_tree_nodevalue=_query_tree_nodevalue,
                         capture_function=n->n.head,
                         node_comparison_yields_true=_capture_on_empty_query_value)
       map(t->t[1:2], r)
6-element Vector{Tuple{Bool, DataStructures.MultiDict{Any, Any}}}:
 (1, DataStructures.MultiDict{Any, Any}(Dict{Any, Vector{Any}}("v2" => [3, 10], "v0" => [1])))
 (0, DataStructures.MultiDict{Any, Any}(Dict{Any, Vector{Any}}()))
 (0, DataStructures.MultiDict{Any, Any}(Dict{Any, Vector{Any}}()))
 (0, DataStructures.MultiDict{Any, Any}(Dict{Any, Vector{Any}}("v2" => [3])))
 (0, DataStructures.MultiDict{Any, Any}(Dict{Any, Vector{Any}}()))
 (0, DataStructures.MultiDict{Any, Any}(Dict{Any, Vector{Any}}()))